requirements.txt
alembic==1.13.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.6.0
beautifulsoup4==4.12.3
blurhash-python==1.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==39.0.2
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.8
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.1
fastapi==0.110.0
filelock==3.13.4
fonttools==4.52.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
Js2Py==0.74
kiwisolver==1.4.5
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib==3.9.0
mysqlclient==2.1.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
orjson==3.9.15
packaging==24.0
pefile==2023.2.7
Pillow==9.4.0
pipenv==2023.12.1
pipwin==0.5.2
platformdirs==4.2.0
playsound==1.3.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pyjsparser==2.7.1
pyparsing==3.1.2
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.9
python3-openid==3.2.0
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.3
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2022.7
tzlocal==5.2
ujson==5.9.0
urllib3==1.26.15
uvicorn==0.28.0
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0